fix(sort): events were not sorting correctly with allday and regular events
When sorting, all day events were at the top of the array causing regular events to be pushed out
incorrectly. Now the function will sort each day individually and then merge the array. Milestone
sorting still needs work as it is not sorting as events on the current day as expected.

fixes #1378, fixes #1247